refactor: Sentry 에러 리포팅 경로 수정 및 관련 파일 삭제
