+++ frontend/src/apis/rest/apiRequest.ts
@@ -1,5 +1,5 @@
 import { ApiError, NetworkError } from './error';
-import { reportApiError } from '@/utils/reportSentryError';
+import { reportApiError } from '@/apis/utils/reportSentryError';
 
 const API_URL = process.env.REACT_APP_API_URL || 'https://api.coffee-shout.com';
