+++ frontend/src/apis/utils/reportSentryError.ts
